Move logging initialisation from dialog to application object.
