+++ src/endless/StringHelperMethods.cpp
@@ -1,0 +1,42 @@
+#include "stdafx.h"
+
+#include "StringHelperMethods.h"
+//#include <atlcomcli.h>
+//#include <atlstr.h>
+
+// utility method for quick char* UTF8 conversion to BSTR
+CComBSTR UTF8ToBSTR(const char *txt) {
+	int wchars_num = MultiByteToWideChar(CP_UTF8, 0, txt, -1, NULL, 0);
+	wchar_t* wstr = new wchar_t[wchars_num];
+	MultiByteToWideChar(CP_UTF8, 0, txt, -1, wstr, wchars_num);
+	CComBSTR return_value(wstr);
+	delete[] wstr;
+
+	return return_value;
+}
+
+CString UTF8ToCString(const char *txt) {
+	int wchars_num = MultiByteToWideChar(CP_UTF8, 0, txt, -1, NULL, 0);
+	wchar_t* wstr = new wchar_t[wchars_num];
+	MultiByteToWideChar(CP_UTF8, 0, txt, -1, wstr, wchars_num);
+	CString return_value(wstr);
+	delete[] wstr;
+
+	return return_value;
+}
+
+CStringA ConvertUnicodeToUTF8(const CStringW& uni)
+{
+	if (uni.IsEmpty()) return ""; // nothing to do
+	CStringA utf8;
+	int cc = 0;
+	// get length (cc) of the new multibyte string excluding the \0 terminator first
+	if ((cc = WideCharToMultiByte(CP_UTF8, 0, uni, -1, NULL, 0, 0, 0) - 1) > 0)
+	{
+		// convert
+		char *buf = utf8.GetBuffer(cc);
+		if (buf) WideCharToMultiByte(CP_UTF8, 0, uni, -1, buf, cc, 0, 0);
+		utf8.ReleaseBuffer();
+	}
+	return utf8;
+}
